Move WolvicContentRendererClient to wolvic/renderer
This is consistent with the rest of the content renderer clients
found in the Chromium sources.
