+++ wolvic/BUILD.gn
@@ -56,12 +56,12 @@ shared_library("libcontent_native") {
     "browser/vr/wvr_thread.h",
     "browser/youtube/youtube_url_loader_request_interceptor.cc",
     "browser/youtube/youtube_url_loader_request_interceptor.h",
+    "renderer/wolvic_content_renderer_client.cc",
+    "renderer/wolvic_content_renderer_client.h",
     "wolvic_browser_context.cc",
     "wolvic_browser_context.h",
     "wolvic_content_browser_client.cc",
     "wolvic_content_browser_client.h",
-    "wolvic_content_renderer_client.cc",
-    "wolvic_content_renderer_client.h",
     "wolvic_content_client.cc",
     "wolvic_content_client.h",
     "wolvic_content_main_delegate.cc",
