+++ wolvic/renderer/wolvic_content_renderer_client.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "wolvic/wolvic_content_renderer_client.h"
+#include "wolvic/renderer/wolvic_content_renderer_client.h"
 
 #include "components/cdm/renderer/key_system_support_update.h"
 #include "components/visitedlink/renderer/visitedlink_reader.h"
